Correcting bugs related to floating point errors, refactoring code
Floating point errors were caused due to discrepancy in the way Python and
Java handle floats. This has not been fully resolved but it has been seen
that it does not affect the result much. Most errors were solved after the
rounding off function was rewritten.

The test file created using getNltkVader.py was not handling non-ascii
values correctly as a result of which java version was giving a different
value. That has been resolved.

While porting the code, the part where idioms are handled was wrongly
implemented in the java version. That has been corrected.

Added tests for comparing results of Amazon Reviews, Movie Reviews and
NYTimes Editorial snippets.